// src/app/interfaces/Artist.ts
// Generated by https://quicktype.io

export interface Artist {
    external_urls: ExternalUrls;
    followers:     Followers;
    genres:        string[];
    href:          string;
    id:            string;
    images:        Image[];
    name:          string;
    popularity:    number;
    type:          string;
    uri:           string;
}

interface ExternalUrls {
    spotify: string;
}

interface Followers {
    href:  null;
    total: number;
}

interface Image {
    height: number | null;
    url:    string;
    width:  number | null;
}